I have added the second aggregate root of infection which has a finder. the finder contains the method find_nearby.
